More work on configuration parsing.
The default was unboundedly recursing - be explicit to stop that.
We also need to do some foolish stuff with token types as only some are ': ToString',
but we need to check string equality on the ones that are.
We also change the structure of the conf parser somewhat; this should make it easier to
actually add in the bits that get out the information we need.