docs: note hex, bin, floats in arithmetic EBNF
Add production rules for hexadecimal and binary integers to stand
alongside the decimal integer.
Ensure that float literals have their own, separate production rules as
integer literals cannot have fraction or exponent parts.
Replace the existing whitespace rule with the EBNF special sequence to
leave how Rust determines whitespace undefined.

I've continued to associate negative numbers with `<dec_integer>`, used
in float as well, which feels like a line call. I'm not sure if
underscores are technically allowed between 0x/0b and the number, but
I'm assuming they probably are. I'm omitting octal as we aren't testing
for it. One interesting side effect of `<float>`'s definition is that
`1E1` is allowed, but it is handled as a floating point, which matches
the tests.