docs: more correct arithmetic EBNF and notes
Finalise the EBNF to be as correct as we care to be without going whole
hog and creating a grammar that guarantees type-correct production -
this would be too long and illegible.

Add some supplementary notes to compensate for this: point out that
literals should be within size and sign bounds, note bitwise operators
only work for integers.